requirements.txt
gradio>=4.3.0
pandas
matplotlib
seaborn
scikit-learn
pillow
groq
